Added sds client and rest road

+++ src/main/scala/com/ubirch/webui/models/graph/DefaultGraphClient.scala
@@ -68,7 +68,7 @@ class DefaultGraphClient(gc: GremlinConnector) extends GraphClient with LazyLogg
               .V()
               .has(Key[String]("hash"), maybeLastHash.head)
               .repeat(_.out("CHAIN"))
-              .limit(n)
+              .limit(n.toLong)
               .count()
               .promise()
             maybeNumberOfUpps.flatMap { numberOfUpps =>
@@ -104,7 +104,7 @@ class DefaultGraphClient(gc: GremlinConnector) extends GraphClient with LazyLogg
         .has(Key[String]("hash"), hash)
         .repeat(_.out("CHAIN"))
         .emit()
-        .limit(n)
+        .limit(n.toLong)
         .path()
         .tail()
         .unfold()
